Change error handling strategy in parameter binding functions
This is done to provide more detailed error message and to prevent continuing
the execution of binding functions chain in case of failure.

Moreover, the query execution process is already has to be wrapped in a
try-catch block as Statement::step() may throw, so there is no point to use
error codes instead of throwing an exception.